requirements.txt
asgiref==3.11.0
Django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.4
